[Spanify] `span` size/offset to conform to `size_t` (#26822)
There's some tightening around values passed into span, with
`StrictNumeric<size_t>` being enforced. This is mostly unimportant when
dealing with compile-time values, as they can be validated on the spot,
however it does affect several places where the value is unknown at
runtime, and a signed value was being used.

Chromium changes:
https://chromium.googlesource.com/chromium/src/+/4937ad9ae64d9d3f05d13c9d041769432a12281b

```
commit 4937ad9ae64d9d3f05d13c9d041769432a12281b
Author: Peter Kasting <pkasting@chromium.org>
Date:   Tue Nov 26 22:58:16 2024 +0000

    Use StrictNumeric<size_t> for more span methods.

    Changes subspan(), split_at(), at(), and get_at() to match pre-existing
    usage in constructors, first(), last(), etc.

    Bug: 364987728
```

Resolves https://github.com/brave/brave-browser/issues/42627

+++ browser/ui/views/tabs/tab_drag_controller.cc
@@ -187,7 +187,8 @@ void TabDragController::DetachAndAttachToNewContext(
     std::vector<tabs::TabHandle> tabs;
     auto* tab_strip_model = browser->tab_strip_model();
     DCHECK_EQ(tab_strip_model, attached_context_->GetTabStripModel());
-    auto drag_data = base::span(drag_data_).subspan(first_tab_index());
+    auto drag_data =
+        base::span(drag_data_).subspan(static_cast<size_t>(first_tab_index()));
     for (const auto& data : drag_data) {
       tabs.push_back(tab_strip_model->GetTabHandleAt(
           tab_strip_model->GetIndexOfWebContents(data.contents)));
